rx: Move iovec filling to the service thread
iovec handling was split between the service and listener thread,
resulting in a very unclear locking model for big chunks of the
call structure. Now that FillReadVec cannot send hard acks, the race
condition that this was designed to prevent no longer exists.

So, move all calls to FillReadVec into the reader thread. This does
mean that the reader may be woken up more often than previously, but
should give us significant benefits in terms of reduced contention
on call->lock

Change-Id: Ibe133c317dace9d8f0d272ca1059b82ded3a58b1