rx: Move service thread variables to own struct
Move all of the elements of the rx_call structure which are already
marked as being exclusively used by the service thread into their own
structure, in an attempt to make the locking rules used around rx_call
more explicit.

Note that there are other elements which are used only by the service
thread, but which aren't currently marked as such. These will be moved
in subsequent patches.

Change-Id: I2f7a1e0298a56f767c1e6c5f11dd2b28e16243df